Fix two Xyce- and ADMS-specific issues
ADMS does not appear to like this usage:

   `define NOT_USED -1e21
   variable=-`NOT_USED;
   if (variable == -`NOT_USED)

That's mainly because it winds up getting expanded  to "--1e21"
and the doubled minus causes problems. Specifically, ADMS pukes while
parsing the verilog-a:

 ./ekv26_SDext_Verilog-A.va: during lexical analysis syntax error at line 208 -- see '-'

This is not specific to Xyce's back-end, as it's happening as the
verilog-A itself is being parsed.

Since NOT_USED is never, ever used anywhere in ekv except as
-`NOT_USED, there's no real reason to carry around the extra unary
minus.

Secondly, Xyce/ADMS is not set up to handle noise contributions except
when they contain a single noise function.  EKV had one contribution
with both  flicker and white noise function in it.  Splitting that one
contribution into two makes Xyce/ADMS happy.